+++ Project_BookStoreCT/Models/ViewModels/BooksIndex_ViewModels.cs
@@ -9,11 +9,11 @@ namespace Project_BookStoreCT.Models.ViewModels
     {
         public int book_id { get; set; }
         public string bookName { get; set; }
-        public int ? price { get; set; }
+        public double ? price { get; set; }
         public int ? quantityExist { get; set; }
         public string image { get; set; }
         public int ? sellNumber { get; set; }
-        public int ? priceSaleOff { get; set; }
+        public double ? priceSaleOff { get; set; }
         public bool ? statusSaleOff { get; set; }
         public string language { get; set; }
         public string categoryName { get; set; }
